+++ app/src/main/java/com/dejssa/radioserver/handlers/StationHandler.java
@@ -2,10 +2,13 @@ package com.dejssa.radioserver.handlers;
 
 import com.dejssa.radioserver.provider.Providers;
 import com.dejssa.radioserver.repository.StationRepository;
+import com.dejssa.radioserver.storage.domain.StationInfo;
 import com.dejssa.radioserver.storage.requests.StationRequest;
 import com.dejssa.radioserver.storage.requests.StationUUIDRequest;
 import com.dejssa.radioserver.storage.responses.StatusResponse;
 import com.google.gson.Gson;
+
+import java.util.ArrayList;
 
 import fi.iki.elonen.NanoHTTPD.IHTTPSession;
 import fi.iki.elonen.NanoHTTPD.Response;
@@ -22,23 +25,23 @@ public class StationHandler {
         this.providers = providers;
     }
 
-    public Response serverSave(IHTTPSession session) {
+    public Response serveSave(IHTTPSession session) {
         StationRequest request = RequestParser.parseObject(session, StationRequest.class);
 
         this.providers.station.save(request.toDomain());
 
-        StatusResponse state = this.providers.radio.state();
+        ArrayList<StationInfo> stations = this.providers.station.find();
 
-        return newFixedLengthResponse(new Gson().toJson(state));
+        return newFixedLengthResponse(new Gson().toJson(stations));
     }
 
-    public Response serverDelete(IHTTPSession session) {
+    public Response serveDelete(IHTTPSession session) {
         StationUUIDRequest request = RequestParser.parseObject(session, StationUUIDRequest.class);
 
         this.providers.station.delete(request.UUID);
 
-        StatusResponse state = this.providers.radio.state();
+        ArrayList<StationInfo> stations = this.providers.station.find();
 
-        return newFixedLengthResponse(new Gson().toJson(state));
+        return newFixedLengthResponse(new Gson().toJson(stations));
     }
 }